After 30 minutes and 34 seconds
This was way too stressful. First I kept getting weird noisy US-ASCII encoding
feedback in the test diffs. I've noticed that happen in recent Ruby versions.

Second I struggled for a bit on the pluralize_bottle conditional order. I know
it's gross but I was trying to finish in 30 minutes and I'm failry satisfied
of that considering this is only a first draft.

Third I had no idea you can't instantiate a non-ascending range. I tried to
do (first..second).to_a a bunch and it wasn't working.

I originally assumed I was running Ruby 2.3 so I tried to use <<~ (strip left
whitespace heredocs) to no avail.

I was surprised to see that the extraction of calls to the verse method, and the
verses method didn't cause too much breakage. I expect that to quickly become
a ball of mud.

Finally I'm pretty sure this is the first time I finish a thing like this
somewhat on time. I'm not sure it's a metric I should be proud of.